Make ZRTP Hello packet parsing more flexible.
I'd eventually like to support the EC25 key agreement option, since
that's likely faster than the mandatory-supported DH3K.  I wrote this
to be fairly inflexible, though, to the point where specifying any
optional arguments will result in a MAC error and cause the handshake
to fail.

This change makes the Hello packet parsing more flexible, so that it
can handle parsing packets with any of the optional parameters
specified.  Once this has been deployed for a long enough period of
time, we could start safely specifying optional parameters and enabling
support for EC25.