Add support for the EC25 KeyAgreement type (ECDHE over P-256).
1) We don't advertise EC25 support in the KA field on the Hello
packet, because it will mess up handshake parsing on older clients
due to a bug.  So we assume that EC25 is supported if it's advertised,
or if the ClientID string matches the updated one.

2) This reduces the DH calculation time by 2x, noticably speeding up
the ZRTP handshake.

3) We have to include "spongy castle," which is just a jarjar renaming
of bouncycastle, because Android removed the ECC support from the BC
provider they ship.