Correctly handle closed socket.
When we receive a signal that terminates the call (e.g. remote
hangup) this causes the RtpSocket.close() method to be called.
If the socket is in a blocking recv() on another thread when this
happens and IOException will be thrown.  In this case we do not
want to propagate this exception up the stack, since this will
cause the 'Ending Call' message to change to 'Client Failed'.